created: first usable version.
GCodeSimpler is a replacement of old GCodeTransfer tool which converts
standard Gcode to obscure proprietary xj3dp format.
The improvement to the former one is focused on UI and output format.
The UI was reformed to a more straight forward drag-and-drop-based one
which avoids unnecessary parameters and makes it easier to use.
The output format was changed to "xj3dp1" to avoid conflict with new
gcode-based xj3dp format (meaningly v2).